+++ src/components/common/Toast/Modal.css
@@ -1,0 +1,36 @@
+.dialog-background {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.3);
+  z-index: 999;
+}
+.dialog-container {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: white;
+  border: 1px solid black;
+  border-radius: 4px;
+  z-index: 9999;
+  width: 280px;
+  height: 110px;
+  display: flex;
+  flex-direction: column;
+}
+.dialog-text {
+  border-bottom: 1px solid gray;
+  font-size: 0.85rem;
+  padding: 10px;
+  margin: 0;
+  flex-grow: 1.7;
+}
+.dialog-button {
+  text-align: center;
+  margin: 0;
+  padding: 10px;
+  cursor: pointer;
+}
