--- src/pages/Mainpage/Mainpage.css
@@ -1,37 +1,0 @@
-.main-container {
-  display: flex;
-  flex-direction: column;
-  width: 70%;
-  background-color: white;
-}
-
-.costTxt {
-  display: flex;
-  padding: 25px;
-  font-size: 1.2rem;
-  font-weight: bold;
-  justify-content: space-between;
-  border: 1px solid #f2f2f2;
-}
-
-#cashReceiptInput {
-  padding: 8px;
-  margin: 10px;
-  border-radius: 5px;
-  border: 1px solid #d9d9d9;
-}
-
-.payBtn {
-  background-color: rgb(255, 235, 0);
-  height: 50px;
-  border: 0px;
-  font-weight: bold;
-  font-size: 1rem;
-  cursor: pointer;
-}
-.toggle-text {
-  display: flex;
-  align-items: center;
-  font-size: 2rem;
-  padding-left: 30px;
-}
